TrackrAPI/ai/requirements.txt
pymongo==4.6.1
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
xgboost==2.1.1
joblib==1.4.2
